requirements.txt
# Core dependencies
gymnasium[classic-control]>=0.29.1
stable-baselines3>=2.3.0
torch>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
moviepy==1.0.3
imageio[ffmpeg]>=2.36.0
scipy>=1.10.0
tensorboard>=2.14.0
pyyaml>=6.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
